Move `ObjectChange` to `core` app
