+++ core/api/urls.py
@@ -10,6 +10,8 @@ router.register("data-files", views.DataFileViewSet)
 router.register("data-sources", views.DataSourceViewSet)
 # Jobs
 router.register("jobs", views.JobViewSet)
+# Change logging
+router.register("object-changes", views.ObjectChangeViewSet)
 
 app_name = "core-api"
 urlpatterns = router.urls
