+++ core/enums.py
@@ -49,3 +49,15 @@ class LogLevel(ChoiceSet):
         (WARNING, "warning", "warning"),
         (FAILURE, "failure", "danger"),
     )
+
+
+class ObjectChangeAction(ChoiceSet):
+    CREATE = "create"
+    UPDATE = "update"
+    DELETE = "delete"
+
+    CHOICES = (
+        (CREATE, "Created", "success"),
+        (UPDATE, "Updated", "warning"),
+        (DELETE, "Deleted", "danger"),
+    )
