+++ core/models/__init__.py
@@ -1,2 +1,3 @@
+from .change_logging import *
 from .data import *
 from .jobs import *
