+++ core/views/data.py
@@ -8,8 +8,22 @@ from peering_manager.views.generic import (
     ObjectView,
 )
 
-from . import filtersets, forms, tables
-from .models import DataFile, DataSource, Job
+from .. import filtersets, forms, tables
+from ..models import DataFile, DataSource
+
+__all__ = (
+    "DataFileListView",
+    "DataFileView",
+    "DataFileDeleteView",
+    "DataFileBulkDeleteView",
+    "DataSourceListView",
+    "DataSourceEditView",
+    "DataSourceBulkEdit",
+    "DataSourceDeleteView",
+    "DataSourceBulkDeleteView",
+    "DataSourceView",
+    "DataSourceFilesView",
+)
 
 
 class DataFileListView(ObjectListView):
@@ -102,29 +116,3 @@ class DataSourceFilesView(ObjectChildrenView):
             .defer("data")
             .prefetch_related("source")
         )
-
-
-class JobListView(ObjectListView):
-    permission_required = "core.view_job"
-    queryset = Job.objects.all()
-    filterset = filtersets.JobFilterSet
-    filterset_form = forms.JobFilterForm
-    table = tables.JobTable
-    template_name = "core/job/list.html"
-
-
-class JobDeleteView(ObjectDeleteView):
-    permission_required = "core.delete_job"
-    queryset = Job.objects.all()
-
-
-class JobBulkDeleteView(BulkDeleteView):
-    permission_required = "core.delete_job"
-    queryset = Job.objects.all()
-    filterset = filtersets.JobFilterSet
-    table = tables.JobTable
-
-
-class JobView(ObjectView):
-    permission_required = "core.view_job"
-    queryset = Job.objects.all()
