+++ extras/api/serializers.py
@@ -4,21 +4,15 @@ from drf_spectacular.utils import extend_schema_field
 from rest_framework import serializers
 
 from extras.utils import FeatureQuery
-from peering_manager.api.fields import ChoiceField, ContentTypeField
-from peering_manager.api.serializers import (
-    BaseModelSerializer,
-    ValidatedModelSerializer,
-)
-from users.api.nested_serializers import NestedUserSerializer
+from peering_manager.api.fields import ContentTypeField
+from peering_manager.api.serializers import ValidatedModelSerializer
 from utils.api import get_serializer_for_model
 
-from ..enums import ObjectChangeAction
 from ..models import (
     IXAPI,
     ConfigContext,
     ConfigContextAssignment,
     ExportTemplate,
-    ObjectChange,
     Tag,
     Webhook,
 )
@@ -28,7 +22,6 @@ __all__ = (
     "ConfigContextSerializer",
     "ConfigContextAssignmentSerializer",
     "ExportTemplateSerializer",
-    "ObjectChangeSerializer",
     "TagSerializer",
     "WebhookSerializer",
     "NestedTagSerializer",
@@ -106,50 +99,6 @@ class IXAPIAccountSerializer(serializers.Serializer):
     url = serializers.CharField()
     api_key = serializers.CharField()
     api_secret = serializers.CharField()
-
-
-class ObjectChangeSerializer(BaseModelSerializer):
-    url = serializers.HyperlinkedIdentityField(
-        view_name="extras-api:objectchange-detail"
-    )
-    user = NestedUserSerializer(read_only=True)
-    action = ChoiceField(choices=ObjectChangeAction, read_only=True)
-    changed_object_type = ContentTypeField(read_only=True)
-    changed_object = serializers.SerializerMethodField(read_only=True)
-
-    class Meta:
-        model = ObjectChange
-        fields = [
-            "id",
-            "url",
-            "display",
-            "time",
-            "user",
-            "user_name",
-            "request_id",
-            "action",
-            "changed_object_type",
-            "changed_object_id",
-            "changed_object",
-            "prechange_data",
-            "postchange_data",
-        ]
-
-    @extend_schema_field(serializers.DictField)
-    def get_changed_object(self, o):
-        """
-        Serialize a nested representation of the changed object.
-        """
-        if o.changed_object is None:
-            return None
-
-        try:
-            serializer = get_serializer_for_model(o.changed_object, prefix="Nested")
-        except Exception:
-            return o.object_repr
-
-        context = {"request": self.context["request"]}
-        return serializer(o.changed_object, context=context).data
 
 
 class TagSerializer(ValidatedModelSerializer):
