+++ extras/api/views.py
@@ -15,7 +15,6 @@ from ..filtersets import (
     ConfigContextAssignmentFilterSet,
     ConfigContextFilterSet,
     ExportTemplateFilterSet,
-    ObjectChangeFilterSet,
     TagFilterSet,
     WebhookFilterSet,
 )
@@ -25,7 +24,6 @@ from ..models import (
     ConfigContext,
     ConfigContextAssignment,
     ExportTemplate,
-    ObjectChange,
     Tag,
     Webhook,
 )
@@ -35,7 +33,6 @@ from .serializers import (
     ExportTemplateSerializer,
     IXAPIAccountSerializer,
     IXAPISerializer,
-    ObjectChangeSerializer,
     TagSerializer,
     WebhookSerializer,
 )
@@ -156,12 +153,6 @@ class IXAPIViewSet(PeeringManagerModelViewSet):
         api.authenticate()
 
         return Response(data=api.accounts.all())
-
-
-class ObjectChangeViewSet(PeeringManagerModelViewSet):
-    queryset = ObjectChange.objects.all()
-    serializer_class = ObjectChangeSerializer
-    filterset_class = ObjectChangeFilterSet
 
 
 class TagViewSet(PeeringManagerModelViewSet):
