+++ extras/enums.py
@@ -26,15 +26,3 @@ class HttpMethod(ChoiceSet):
         (PATCH, "PATCH"),
         (DELETE, "DELETE"),
     )
-
-
-class ObjectChangeAction(ChoiceSet):
-    CREATE = "create"
-    UPDATE = "update"
-    DELETE = "delete"
-
-    CHOICES = (
-        (CREATE, "Created", "success"),
-        (UPDATE, "Updated", "warning"),
-        (DELETE, "Deleted", "danger"),
-    )
