+++ extras/models/__init__.py
@@ -1,4 +1,3 @@
-from .change_logging import *
 from .configcontext import *
 from .ixapi import *
 from .models import *
