+++ extras/views.py
@@ -1,9 +1,7 @@
 from django.contrib.contenttypes.models import ContentType
 from django.db.models import Count
-from django.shortcuts import get_object_or_404, render
-from django.views.generic import View
+from django.shortcuts import get_object_or_404
 
-from extras.enums import ObjectChangeAction
 from peering.models import InternetExchange
 from peering_manager.views.generic import (
     BulkDeleteView,
@@ -12,15 +10,12 @@ from peering_manager.views.generic import (
     ObjectEditView,
     ObjectListView,
     ObjectView,
-    PermissionRequiredMixin,
 )
-from utils.functions import shallow_compare_dict
 
 from .filtersets import (
     ConfigContextFilterSet,
     ExportTemplateFilterSet,
     IXAPIFilterSet,
-    ObjectChangeFilterSet,
     TagFilterSet,
     WebhookFilterSet,
 )
@@ -32,7 +27,6 @@ from .forms import (
     ExportTemplateForm,
     IXAPIFilterForm,
     IXAPIForm,
-    ObjectChangeFilterForm,
     TagBulkEditForm,
     TagFilterForm,
     TagForm,
@@ -44,7 +38,6 @@ from .models import (
     ConfigContext,
     ConfigContextAssignment,
     ExportTemplate,
-    ObjectChange,
     Tag,
     TaggedItem,
     Webhook,
@@ -54,7 +47,6 @@ from .tables import (
     ConfigContextTable,
     ExportTemplateTable,
     IXAPITable,
-    ObjectChangeTable,
     TaggedItemTable,
     TagTable,
     WebhookTable,
@@ -229,83 +221,6 @@ class IXAPIEditView(ObjectEditView):
 class IXAPIDeleteView(ObjectDeleteView):
     permission_required = "extras.delete_ixapi"
     queryset = IXAPI.objects.all()
-
-
-class ObjectChangeList(ObjectListView):
-    permission_required = "extras.view_objectchange"
-    queryset = ObjectChange.objects.select_related("user").prefetch_related(
-        "changed_object"
-    )
-    filterset = ObjectChangeFilterSet
-    filterset_form = ObjectChangeFilterForm
-    table = ObjectChangeTable
-    template_name = "extras/object_change/list.html"
-
-
-class ObjectChangeView(PermissionRequiredMixin, View):
-    permission_required = "extras.view_objectchange"
-
-    def get(self, request, pk):
-        instance = get_object_or_404(ObjectChange, pk=pk)
-
-        related_changes = ObjectChange.objects.filter(
-            request_id=instance.request_id
-        ).exclude(pk=instance.pk)
-        related_changes_table = ObjectChangeTable(
-            data=related_changes[:50], orderable=False
-        )
-
-        object_changes = ObjectChange.objects.filter(
-            changed_object_type=instance.changed_object_type,
-            changed_object_id=instance.changed_object_id,
-        )
-
-        next_change = (
-            object_changes.filter(time__gt=instance.time).order_by("time").first()
-        )
-        previous_change = (
-            object_changes.filter(time__lt=instance.time).order_by("-time").first()
-        )
-
-        if (
-            not instance.prechange_data
-            and instance.action
-            in [ObjectChangeAction.UPDATE, ObjectChangeAction.DELETE]
-            and previous_change
-        ):
-            non_atomic_change = True
-            prechange_data = previous_change.postchange_data
-        else:
-            non_atomic_change = False
-            prechange_data = instance.prechange_data
-
-        if prechange_data and instance.postchange_data:
-            diff_added = shallow_compare_dict(
-                prechange_data or {},
-                instance.postchange_data or {},
-                exclude=["updated"],
-            )
-            diff_removed = (
-                {x: prechange_data.get(x) for x in diff_added} if prechange_data else {}
-            )
-        else:
-            diff_added = None
-            diff_removed = None
-
-        return render(
-            request,
-            "extras/object_change/details.html",
-            {
-                "instance": instance,
-                "diff_added": diff_added,
-                "diff_removed": diff_removed,
-                "next_change": next_change,
-                "previous_change": previous_change,
-                "related_changes_table": related_changes_table,
-                "related_changes_count": related_changes.count(),
-                "non_atomic_change": non_atomic_change,
-            },
-        )
 
 
 class TagList(ObjectListView):
