+++ peering_manager/middleware.py
@@ -12,7 +12,7 @@ from django.db import ProgrammingError
 from django.http import Http404, HttpResponseRedirect
 from django.urls import reverse
 
-from extras.context_managers import change_logging
+from core.context_managers import change_logging
 from utils.api import is_api_request, rest_api_server_error
 
 from .views import handler_500
