+++ templates/_base.html
@@ -168,7 +168,7 @@
               <a href="{% url 'extras:webhook_list' %}" class="list-group-item bg-dark text-white{% if request.path|contains:'/webhooks/' %} active{% endif %}">
                 <span class="ml-4">Webhooks</span>
               </a>
-              <a href="{% url 'extras:objectchange_list' %}" class="list-group-item bg-dark text-white{% if request.path|contains:'/extras/changelog/' %} active{% endif %}">
+              <a href="{% url 'core:objectchange_list' %}" class="list-group-item bg-dark text-white{% if request.path|contains:'/extras/changelog/' %} active{% endif %}">
                 <span class="ml-4">Change Log</span>
               </a>
             </div>
